Move tests to PublicOffer
